requirements.txt
streamlit
pinecone-client
sentence-transformers
langchain
requests
